half solve

+++ src/main/java/com/songr/songr/repository/AlbumsRepository.java
@@ -4,5 +4,8 @@ import com.songr.songr.model.Album;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 
-public interface AlbumsRepository extends CrudRepository<Album, Long> {
+import java.util.Map;
+
+public interface AlbumsRepository extends CrudRepository<Album, Integer> {
+
 }
